refactor openquake support into psha_adapter package;
