--- nzshm_model/nrml/__init__.py
@@ -1,3 +1,0 @@
-"""
-"""
-# from .logic_tree import NrmlDocument
