+++ nzshm_model/psha_adapter/openquake/__init__.py
@@ -1,0 +1,4 @@
+"""
+"""
+from .logic_tree import NrmlDocument
+from .simple_nrml import OpenquakeSimplePshaAdapter
